Move rust transcation-builder into client/
Closes: #5801
