+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "client/json-rpc",
     "client/libra-dev",
     "client/swiss-knife",
+    "client/transaction-builder",
     "common/bitvec",
     "common/bounded-executor",
     "common/channel",
@@ -92,7 +93,6 @@ members = [
     "language/tools/transaction-replay",
     "language/tools/vm-genesis",
     "language/transaction-builder",
-    "language/transaction-builder/generated",
     "language/transaction-builder/generator",
     "language/vm",
     "language/vm/serializer-tests",
