+++ client/transaction-builder/Cargo.toml
@@ -12,13 +12,15 @@ edition = "2018"
 once_cell = "1.4.1"
 serde = { version = "1.0.115", features = ["derive"] }
 
-move-core-types = { path = "../../move-core/types", version = "0.1.0" }
-libra-types = { path = "../../../types", version = "0.1.0" }
-libra-workspace-hack = { path = "../../../common/workspace-hack", version = "0.1.0" }
+compiled-stdlib = { path = "../../language/stdlib/compiled",  version = "0.1.0" }
+lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
+move-core-types = { path = "../../language/move-core/types", version = "0.1.0" }
+libra-types = { path = "../../types", version = "0.1.0" }
+libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 
 proptest = { version = "0.10.1", optional = true }
 proptest-derive = { version = "0.2.0", optional = true }
-libra-proptest-helpers = { path = "../../../common/proptest-helpers", version = "0.1.0", optional = true }
+libra-proptest-helpers = { path = "../../common/proptest-helpers", version = "0.1.0", optional = true }
 
 [features]
 default = []
