+++ language/e2e-testsuite/Cargo.toml
@@ -19,7 +19,7 @@ libra-types = { path = "../../types", version = "0.1.0", features = ["fuzzing"] 
 move-core-types = { path = "../move-core/types", version = "0.1.0" }
 move-vm-types = { path = "../move-vm/types", version = "0.1.0" }
 transaction-builder = { path = "../transaction-builder", version = "0.1.0"}
-transaction-builder-generated = { path = "../transaction-builder/generated", version = "0.1.0"}
+transaction-builder-generated = { path = "../../client/transaction-builder", version = "0.1.0"}
 vm = { path = "../vm", version = "0.1.0" }
 libra-vm = { path = "../libra-vm", version = "0.1.0" }
 proptest = "0.10.1"
