+++ secure/key-manager/Cargo.toml
@@ -25,7 +25,7 @@ libra-secure-storage = { path = "../../secure/storage", version = "0.1.0" }
 libra-secure-time = { path = "../../secure/time", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
-transaction-builder-generated = { path = "../../language/transaction-builder/generated", version = "0.1.0" }
+transaction-builder-generated = { path = "../../client/transaction-builder", version = "0.1.0" }
 
 [dev-dependencies]
 anyhow = "1.0.32"
