+++ summaries/summary-full.toml
@@ -877,7 +877,7 @@ features = ['default']
 [[target-package]]
 name = 'transaction-builder-generated'
 version = '0.1.0'
-workspace-path = 'language/transaction-builder/generated'
+workspace-path = 'client/transaction-builder'
 status = 'initial'
 features = ['default', 'fuzzing', 'libra-proptest-helpers', 'proptest', 'proptest-derive']
